Port more of env_init to Rust
This does not adopt it, and might performs the initialization partially double.

confstr and `_CS_PATH` are not available with Rust libc, and are instead used
over FFI.

I am uncertain about the porting status of `inheriteds`, it appears to not yet
have a Rust-equivalent. That should probably be added.